chore: 🤖 check helm.pgp exists

+++ scripts/install-apt.sh
@@ -16,7 +16,9 @@ fi
 echo "deb [signed-by=/usr/share/keyrings/hashicorp-archive-keyring.gpg] https://apt.releases.hashicorp.com $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/hashicorp.list
 
 # Helm
-curl https://baltocdn.com/helm/signing.asc | sudo gpg --batch --yes --dearmor -o /usr/share/keyrings/helm.gpg
+if [ ! -f /usr/share/keyrings/helm.gpg ]; then
+  curl https://baltocdn.com/helm/signing.asc | sudo gpg --batch --yes --dearmor -o /usr/share/keyrings/helm.gpg
+fi
 echo "deb [signed-by=/usr/share/keyrings/helm.gpg] https://baltocdn.com/helm/stable/debian/ all main" \
   | sudo tee /etc/apt/sources.list.d/helm-stable-debian.list > /dev/null
